Started testing again

+++ db/migrate/20140404221253_drop_text_table.rb
@@ -1,6 +1,6 @@
 class DropTextTable < ActiveRecord::Migration
   def up
-    drop_table :texts
+    drop_table :texts if table_exists? :texts
   end
 
   def down
