+++ db/migrate/20140404221343_drop_page_titles_table.rb
@@ -1,6 +1,6 @@
 class DropPageTitlesTable < ActiveRecord::Migration
   def up
-    drop_table :page_titles
+    drop_table :page_titles if table_exists? :page_titles
   end
 
   def down
