+++ db/migrate/20140423182027_drop_public_themes_table.rb
@@ -1,6 +1,6 @@
 class DropPublicThemesTable < ActiveRecord::Migration
 	def up
-		drop_table :public_themes
+		drop_table :public_themes if table_exists? :public_themes
 	end
 
 	def down
