+++ db/migrate/20140427151723_drop_public_themes_and_public_code_files_talbe.rb
@@ -1,6 +1,6 @@
 class DropPublicThemesAndPublicCodeFilesTalbe < ActiveRecord::Migration
 	def up
-		drop_table :public_themes
+		drop_table :public_themes if table_exists? :public_themes
 	end
 
 	def down
